Use Data Manager to store save states. Also tidied up folders.
