+++ lostphone/js/libs/Common.js
@@ -1,7 +1,4 @@
 // --- Common
-// --------------------------------------------------------
-// Define some bunch of methods commons to all scenes
-
 Object.defineProperty(Phaser.Structs.List.prototype, 'current', {
     get() {
         return this.getAt(this.position);
