+++ lostphone/js/libs/GameSettings.js
@@ -1,69 +1,97 @@
-import { PhoneEvents } from '/scenes/Bootstrap';
+import { PhoneEvents } from '/scenes/main/Bootstrap';
 export default class GameSettings {
-    constructor(game) {
-        this.game = game;
-        this.options = {
-            'muteSound': {
-                'gameSetting': 'sound.mute',
-                'defaultValue': false
-            },
-            'notificationPopup': {
-                'defaultValue': false
-            }
-        };
+  constructor(game) {
+    this.game = game;
+    this.options = {
+      'muteSound': {
+          'gameSetting': 'sound.mute',
+          'defaultValue': false
+      },
+      'notificationPopup': {
+          'defaultValue': false
+      }
+    };
 
-        for (let key in this.options) {
-            this.game.state['settings'][key] = this.options[key]['defaultValue'];
-        }
+    let settings = this.game.registry.get('settings');
+    for (let key in this.options) {
+      settings[key] = this.options[key]['defaultValue'];
+      this.game.registry.set('settings', settings);
+    }
+  }
+
+  /**
+   * Toggle a game setting
+   * @param {*} key
+   */
+  toggleSetting(key)
+  {
+    let t = this;
+    t.setSettingValue(key, !t.getSettingValue(key));
+  }
+
+  /**
+   * Set a value to a setting
+   * @param {*} key
+   * @param {*} value
+   */
+  setSettingValue(key, value)
+  {
+    let t = this;
+    let settings = this.game.registry.get('settings');
+
+    if (key in t.options) {
+      settings[key] = value;
+      this.game.registry.set('settings', settings);
+
+      // Update game setting (if any)
+      t.setGameConfigValue(key);
+      t.game.events.emit(PhoneEvents.SettingsUpdated);
+      t.game.save('autosave');
+    }
+  }
+
+  /**
+   * Change a Phaser game setting with the stored value
+   * @param {*} key
+   */
+  setGameConfigValue(key) {
+    let t = this;
+    let settings = this.game.registry.get('settings');
+
+    if (t.options[key]['gameSetting'] !== undefined) {
+      setPropertyInPath(
+        t.options[key]['gameSetting'],
+        t.game,
+        settings[key]
+      );
+    }
+  }
+
+  /**
+   * Get a setting value
+   * @param {*} key
+   */
+  getSettingValue(key)
+  {
+    let t = this;
+    let settings = this.game.registry.get('settings');
+
+    if (key in t.options) {
+      return settings[key];
     }
 
-    toggleSetting(key)
-    {
-        let t = this;
-        t.setSettingValue(key, !t.getSettingValue(key));
+    return undefined;
+  }
+
+  /**
+   * Sync all PhaserGame settings
+   */
+  fullSync()
+  {
+    let t = this;
+
+    for (let key in t.options) {
+      t.setGameConfigValue(key);
     }
-
-    setSettingValue(key, value)
-    {
-        let t = this;
-        if (key in t.options) {
-            t.game.state['settings'][key] = value;
-
-            // Update game setting (if any)
-            t.setGameConfigValue(key);
-            t.game.events.emit(PhoneEvents.SettingsUpdated);
-            t.game.save('autosave');
-        }
-    }
-
-    setGameConfigValue(key) {
-        let t = this;
-
-        if (t.options[key]['gameSetting'] !== undefined) {
-            setPropertyInPath(
-                t.options[key]['gameSetting'],
-                t.game,
-                t.game.state['settings'][key]
-            );
-        }
-    }
-
-    getSettingValue(key)
-    {
-        let t = this;
-        if (key in t.options) {
-            return t.game.state['settings'][key];
-        }
-
-        return undefined;
-    }
-
-    fullSync()
-    {
-        let t = this;
-
-        for (let key in t.options) {
-            t.setGameConfigValue(key);
-        }
-    }
+  }
 }
