+++ lostphone/js/libs/dataManager/ExtendPhaserDataManager.js
@@ -1,0 +1,17 @@
+/**
+ * Serialize method
+ */
+Phaser.Data.DataManager.prototype.serialize = function() {
+    return JSON.stringify(this.getAll());
+};
+
+/**
+ * Unserialize method
+ */
+Phaser.Data.DataManager.prototype.unserialize = function(saveObject) {
+    let values = JSON.parse(saveObject);
+    for (key in values) {
+        this.set(key, values[key]);
+    }
+    return true;
+};
