+++ lostphone/js/libs/scene/ExtendPhaserScene.js
@@ -1,0 +1,12 @@
+// Phaser.Scene
+Phaser.Scene.prototype.log = function(message) {
+    message = '[Scene:' + this.scene.key + '] ' + message;
+
+    if (this.game.debug) {
+        console.log(message);
+    }
+}
+
+Phaser.Scene.prototype.calcDPR = function(elem) {
+    return Math.floor(elem * this.assetsDPR);
+}
