--- lostphone/js/plugins/README.txt
@@ -1,5 +1,0 @@
-PLUGINS
-=======
-
-* phaser-plugin-scene-watcher: Scene monitoring and debugging for Phaser 3 (https://github.com/samme/phaser-plugin-scene-watcher). Data downloaded: 26/5/2020
-
