+++ lostphone/js/scenes/main/PhoneApp.js
